+++ app/templates/page.tsx
@@ -1,5 +1,5 @@
 import React from "react"
-import HighLightCodeBoard from "../ui/cssTemplates/highlightCodeBoard";
+import HighLightCodeBoard from "../ui/css-templates/highlightCodeBoard";
 import Footer from "../ui/footer";
 import Header from "../ui/header";
 import BaseTemplate from "../ui/baseTemplate";
@@ -7,23 +7,28 @@ import MainTemplate from "../ui/mainTemplate";
 
 const CssTemplates = () => {
 
-    const headerFooter = `<div className="flex flex-col h-screen justify-between bg-gradient-to-b from-black/90 to-slate-900 text-base text-slate-100">
-    <header className="mx-20 text-slate-100 p-4 border-b border-slate-50/20">
-        <h1>Tailwind CSS - code example </h1>
+    const headerFooter = `<div className="flex flex-col min-h-screen justify-between text-sm text-slate-100">
+    <header className="text-slate-100 p-4 bg-gray-600">
+        <h1 className='font-semibold'>Header section</h1>
     </header>
-    <main className="mb-auto p-4 text-slate-100 text-base">
-        <div className="p-2">
-            <div className="border-b border-teal-200 w-fit">
-                <h1>Some examples with CSS:</h1>
-            </div>
+    <main className="mb-auto p-4 text-slate-100 bg-gray-800 h-screen">
+        <div className="p-2 w-fit">
+            <h1 className='font-semibold'>Title 1</h1>
+            <p>Lorem ipsum dolor sit amet consectetur adipisicing elit. Illo blanditiis quam nulla officia eaque.</p>
+        </div>
+
+        <div className="p-2 w-fit">
+            <h1 className='font-semibold'>Title 2</h1>
+            <p>Lorem ipsum dolor sit amet consectetur adipisicing elit.</p>
         </div>
     </main>
-    <footer className="mx-20 h-20 text-slate-100 p-4">
-        <p>&copy; 2023 Mon Site Web. Tous droits réservés.</p>
+    <footer className="text-slate-100 p-4 bg-gray-600">
+        <h1 className='font-semibold'>Footer section</h1>
+        <p>&copy; 2023 My Web site. All rights reserved.</p>
     </footer>
 </div>`
 
-    const headerFooterRendering = headerFooter.replace('h-screen', 'h-full')
+    const headerFooterRendering = headerFooter.replace('min-h-screen', 'h-full')
     const headerFooterHeight = 'h-[500px]'
 
     const button = `<button className='rounded p-4 text-slate-100 bg-blue-600 text-sm'>My Button</button>`