+++ app/ui/baseTemplate.tsx
@@ -1,12 +1,26 @@
+'use client'
+
 import React from 'react'
+import { useThemeMode } from '../context'
 
 
 export default function BaseTemplate({ children }: { children: React.ReactNode }) {
+    const {themeMode} = useThemeMode()
+    const _style = {
+        bg: 'dark:bg-black/90 '
+    }
+    
     return (
-        <div className={`flex flex-col h-screen justify-between 
-        bg-gradient-to-b from-gray-950 to-slate-900
-        text-base text-slate-100`}>
-            {children}
+        <div className={themeMode}>
+            <div className={`flex flex-col min-h-screen justify-between 
+                text-base bg-gradient-to-b 
+                dark:from-gray-950 dark:to-slate-900
+                from-gray-300 to-gray-100
+                text-neutral-800 dark:text-slate-100
+                overflow-x-auto 
+                `}>
+                {children}
+            </div>
         </div>
     )
 }