+++ app/ui/css-templates/highlighted.tsx
@@ -2,7 +2,7 @@
 
 import React, { useState } from "react"
 import { Prism as SyntaxHighlighter } from 'react-syntax-highlighter';
-import { atomDark } from 'react-syntax-highlighter/dist/esm/styles/prism';
+import { atomDark, vscDarkPlus, a11yDark, dracula, duotoneDark, duotoneSea } from 'react-syntax-highlighter/dist/esm/styles/prism';
 import { CopyToClipboard } from 'react-copy-to-clipboard';
 import { 
     ClipboardDocumentCheckIcon, 
@@ -10,14 +10,19 @@ import {
 } from "@heroicons/react/20/solid";
 
 
+const STYLE = duotoneSea
+
+
 const Highlighted = ({
     title,
     height,
-    codeString
+    codeString,
+    language
 }: {
     title: string,
     height: string,
-    codeString: string
+    codeString: string,
+    language: string
 }) => {
     const [copyIcon, setCopyIcon] = useState(<DocumentDuplicateIcon className={`h-5 w-5`}/>)
     const [copyText, setCopyText] = useState('Copy')
@@ -27,10 +32,13 @@ const Highlighted = ({
     }
 
     return (
-        <div className={`shadow-2xl shadow-gray-600 ${height} w-full max-w-[700px] p-2 bg-gray-900 rounded-md border border-gray-500 m-1`}>
+        <div className={`overflow-y-auto shadow-sm shadow-teal-400 ${height} w-full max-w-[700px] p-2 bg-gray-900 rounded-md border border-gray-500 m-1`}>
             
             <div className={`flex justify-between text-gray-400 p-2 border-b border-gray-500`}>
-                <div>{title}</div>
+                <div className="flex w-full">
+                    <div>{title}</div> 
+                    <div className="ml-10 italic">Language: {language}</div>
+                </div>
                 <CopyToClipboard text={codeString} onCopy={onCopy}>
                 <button className="flex">
                     <span>{copyIcon}</span>
@@ -40,8 +48,8 @@ const Highlighted = ({
             </div>
 
             <SyntaxHighlighter 
-                language="xml" 
-                style={atomDark} 
+                language={language}
+                style={STYLE} 
                 lineNumberContainerStyle={{
                     color: 'rgb(156 163 175)',
                     backgroundColor: '#f0f0f0',