+++ app/ui/mainTemplate.tsx
@@ -3,7 +3,7 @@ import React from 'react'
 
 export default function MainTemplate({ children }: { children: React.ReactNode }) {
     return (
-        <div className={`mb-auto p-4 text-slate-100 text-base`}>
+        <div className={`mb-auto p-4 text-base overflow-x-auto `}>
             {children}
         </div>
     )