+++ tailwind.config.ts
@@ -8,6 +8,34 @@ const config: Config = {
   ],
   theme: {
     extend: {
+      colors: {
+        darkest: {
+          50: '#fdf8f6',
+          100: '#f2e8e5',
+          200: '#eaddd7',
+          300: '#e0cec7',
+          400: '#d2bab0',
+          500: '#bfa094',
+          600: '#a18072',
+          700: '#977669',
+          800: '#846358',
+          900: '#010101',
+          950: '#010101',
+        },
+          manta: {
+            50: "#000000",
+            100: "#1a1a1a",
+            200: "#333333",
+            300: "#4d4d4d",
+            400: "#666666",
+            500: "#808080",
+            600: "#999999",
+            700: "#b3b3b3",
+            800: "#cccccc",
+            900: "#e6e6e6",
+            950: "#ffffff"
+          },
+      },
       backgroundImage: {
         'gradient-radial': 'radial-gradient(var(--tw-gradient-stops))',
         'gradient-conic':
@@ -16,5 +44,6 @@ const config: Config = {
     },
   },
   plugins: [],
+  darkMode: 'class',
 }
 export default config