+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 ```
 # Changelog
 
+### feat: dark mode, shade of color, highlight code
+
+- Add React Syntax Highlighter
+- page for highlighting code (header, footer, ...)
+- page to choose shade of color for tailwindcss
+
+---
+
 
 ### feat: Init repo